+++ services/web/server/src/simcore_service_webserver/db/base_repository.py
@@ -2,12 +2,11 @@
 # TODO: a repo: some members acquire and retrieve connection
 # TODO: a repo: any validation error in a repo is due to corrupt data in db!
 
-from typing import Optional
 
 from aiohttp import web
 from aiopg.sa.engine import Engine
 
-from ._constants import APP_DB_ENGINE_KEY, RQT_USERID_KEY
+from .._constants import APP_DB_ENGINE_KEY, RQT_USERID_KEY
 
 
 class BaseRepository:
@@ -20,7 +19,7 @@ class BaseRepository:
         #  user_id, product_name = request[RQT_USERID_KEY], request[RQ_PRODUCT_KEY]
 
         self._engine: Engine = request.app[APP_DB_ENGINE_KEY]
-        self._user_id: Optional[int] = request.get(RQT_USERID_KEY)
+        self._user_id: int | None = request.get(RQT_USERID_KEY)
 
         assert isinstance(self._engine, Engine)  # nosec
 
@@ -29,5 +28,5 @@ class BaseRepository:
         return self._engine
 
     @property
-    def user_id(self) -> Optional[int]:
+    def user_id(self) -> int | None:
         return self._user_id
