+++ services/web/server/src/simcore_service_webserver/garbage_collector_utils.py
@@ -6,7 +6,7 @@ from aiopg.sa.result import RowProxy
 from models_library.users import GroupID, UserID
 from simcore_postgres_database.errors import DatabaseError
 
-from .db_models import GroupType
+from .db.db_models import GroupType
 from .groups.api import get_group_from_gid
 from .projects.db import APP_PROJECT_DBAPI, ProjectAccessRights
 from .projects.exceptions import ProjectNotFoundError
