+++ services/web/server/src/simcore_service_webserver/invitations/plugin.py
@@ -8,7 +8,7 @@ from aiohttp import web
 from servicelib.aiohttp.application_setup import ModuleCategory, app_module_setup
 
 from .._constants import APP_SETTINGS_KEY
-from ..db import setup_db
+from ..db.plugin import setup_db
 from ._client import invitations_service_api_cleanup_ctx
 from ._core import (
     extract_invitation,
