+++ services/web/server/src/simcore_service_webserver/products/_db.py
@@ -6,8 +6,8 @@ from aiopg.sa.connection import SAConnection
 from aiopg.sa.result import ResultProxy, RowProxy
 from simcore_postgres_database.models.products import jinja2_templates
 
-from ..db_base_repository import BaseRepository
-from ..db_models import products
+from ..db.base_repository import BaseRepository
+from ..db.db_models import products
 from ._model import Product
 
 log = logging.getLogger(__name__)
