+++ services/web/server/src/simcore_service_webserver/users/api.py
@@ -15,8 +15,8 @@ from aiopg.sa.result import RowProxy
 from models_library.users import GroupID, UserID
 from simcore_postgres_database.models.users import UserNameConverter, UserRole
 
-from ..db import get_database_engine
-from ..db_models import GroupType, groups, user_to_groups, users
+from ..db.db_models import GroupType, groups, user_to_groups, users
+from ..db.plugin import get_database_engine
 from ..groups.schemas import convert_groups_db_to_schema
 from ..login.storage import AsyncpgStorage, get_plugin_storage
 from ..security.api import clean_auth_policy_cache
