+++ services/web/server/tests/conftest.py
@@ -23,7 +23,7 @@ from pytest_simcore.simcore_webserver_projects_rest_api import NEW_PROJECT
 from servicelib.aiohttp.long_running_tasks.server import TaskStatus
 from servicelib.json_serialization import json_dumps
 from simcore_service_webserver.application_settings_utils import convert_to_environ_vars
-from simcore_service_webserver.db_models import UserRole
+from simcore_service_webserver.db.db_models import UserRole
 from simcore_service_webserver.projects._crud_create_utils import (
     OVERRIDABLE_DOCUMENT_KEYS,
 )
