+++ services/web/server/tests/integration/01/test_exporter_requests_handlers.py
@@ -37,7 +37,7 @@ from simcore_service_webserver.application import (
     setup_users,
 )
 from simcore_service_webserver.application_settings import setup_settings
-from simcore_service_webserver.db import setup_db
+from simcore_service_webserver.db.plugin import setup_db
 from simcore_service_webserver.exporter import settings as exporter_settings
 from simcore_service_webserver.exporter._formatter.archive import get_sds_archive_path
 from simcore_service_webserver.projects.models import ProjectDict
