+++ services/web/server/tests/integration/01/test_garbage_collection.py
@@ -29,8 +29,8 @@ from settings_library.redis import RedisDatabase, RedisSettings
 from simcore_postgres_database.models.users import UserRole
 from simcore_service_webserver import garbage_collector_core
 from simcore_service_webserver.application_settings import setup_settings
-from simcore_service_webserver.db import setup_db
-from simcore_service_webserver.db_models import projects, users
+from simcore_service_webserver.db.db_models import projects, users
+from simcore_service_webserver.db.plugin import setup_db
 from simcore_service_webserver.director.plugin import setup_director
 from simcore_service_webserver.director_v2.plugin import setup_director_v2
 from simcore_service_webserver.garbage_collector import setup_garbage_collector
