+++ services/web/server/tests/integration/02/notifications/test_rabbitmq_consumers.py
@@ -42,7 +42,7 @@ from settings_library.rabbit import RabbitSettings
 from simcore_postgres_database.models.projects import projects
 from simcore_postgres_database.models.users import UserRole
 from simcore_service_webserver.application_settings import setup_settings
-from simcore_service_webserver.db import setup_db
+from simcore_service_webserver.db.plugin import setup_db
 from simcore_service_webserver.diagnostics.plugin import setup_diagnostics
 from simcore_service_webserver.director_v2.plugin import setup_director_v2
 from simcore_service_webserver.login.plugin import setup_login
