+++ services/web/server/tests/unit/with_dbs/01/test_groups.py
@@ -25,7 +25,7 @@ from servicelib.aiohttp.application import create_safe_application
 from simcore_postgres_database.models.users import UserRole
 from simcore_service_webserver._meta import API_VTAG
 from simcore_service_webserver.application_settings import setup_settings
-from simcore_service_webserver.db import setup_db
+from simcore_service_webserver.db.plugin import setup_db
 from simcore_service_webserver.groups import _handlers
 from simcore_service_webserver.groups._db import (
     _DEFAULT_GROUP_OWNER_ACCESS_RIGHTS,
