+++ services/web/server/tests/unit/with_dbs/02/test_projects_handlers__list_search.py
@@ -22,7 +22,7 @@ from pytest_simcore.helpers.utils_webserver_unit_with_db import (
     standard_role_response,
 )
 from simcore_service_webserver._meta import api_version_prefix
-from simcore_service_webserver.db_models import UserRole
+from simcore_service_webserver.db.db_models import UserRole
 from simcore_service_webserver.projects.models import ProjectDict
